Wrap the Alert web component to parse the `hideIcon` boolean
