+++ packages/design-system/src/components/Alert/index.ts
@@ -1,1 +1,1 @@
-export { default as Alert } from './Alert';
+export * from './Alert';
