+++ packages/design-system/src/components/web-components/ds-alert.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { define } from './preactement/define';
+import { Alert, AlertProps } from '../Alert';
+
+const attributes = [
+  'class-name',
+  'heading',
+  'heading-id',
+  'hide-icon',
+  'role',
+  'weight',
+  'variation',
+  'analytics',
+  'analytics-label-override',
+  'analytics-event-type-override',
+  'analytics-parent-heading',
+  'analytics-parent-type',
+  'id',
+];
+
+interface WrapperProps extends Omit<AlertProps, 'hideIcon'> {
+  hideIcon?: string;
+}
+
+const Wrapper = ({ hideIcon, ...otherProps }: WrapperProps) => (
+  <Alert {...otherProps} hideIcon={hideIcon && Boolean(JSON.parse(hideIcon))} />
+);
+
+define('ds-alert', () => Wrapper, { attributes });
